requirements.txt
flask==2.3.2
flask-sqlalchemy==3.1.1
flask-wtf==1.2.2
python-dotenv==1.0.0
flask-login==0.6.2
pandas==2.2.2
reportlab==4.1.0
flask-mail==0.9.1
flask-migrate==4.1.0
sqlalchemy==2.0.41
alembic==1.15.2
itsdangerous==2.2.0
werkzeug==3.1.3
wtforms==3.2.1
openpyxl==3.1.5